Refactor: Move Proxy methods from server.go to proxy.go
This change improves code organization by consolidating all Proxy-related methods in proxy.go, reducing server.go size from 1193 to 682 lines.

Moved methods:
- Connection management: connectToUpstreamServer, connectToUpstreamServers, ConnectToUpstreams
- AMQP protocol: handshake, runHeartbeat, shutdown, sendConnectionError
- Message processing: readClientFrame, process, dispatchN, dispatch0, send, recv
- Utilities: sortNodesByLeastConnections

Benefits:
- Better separation of concerns between Server and Proxy responsibilities
- Improved code organization and maintainability
- All Proxy functionality now consolidated in a single file
